requirements.txt
ping3
speedtest-cli
psutil
geopy
gps
netifaces
matplotlib